Refactor layout machine
